backend/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "eva-backend"
version = "0.1.0"
description = "EVA Medical Assistant — Flask API (POC)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "flask>=3.0,<4",
  "redis>=5.0,<6",
  "python-dotenv>=1.0,<2",
  "gunicorn>=22.0,<24",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["eva_backend*"]
